+++ Sources/SyntaxAnalysis/TypeSyntaxInspector.swift
@@ -55,7 +55,8 @@ struct TypeSyntaxInspector {
         } else if let funcType = typeSyntax.as(FunctionTypeSyntax.self) {
             // Function type.
             let argumentTypes = funcType.parameters.flatMapSet { types(for: $0.type) }
-            return types(for: funcType.returnClause.type).union(argumentTypes)
+            let throwTypes = funcType.effectSpecifiers?.throwsClause?.type.map { types(for: $0) } ?? []
+            return types(for: funcType.returnClause.type).union(argumentTypes).union(throwTypes)
         } else if let arrayType = typeSyntax.as(ArrayTypeSyntax.self) {
             // Array type.
             return types(for: arrayType.element)